RR-211 - Refactor router request handlers into a single module
