+++ server/routes/routerRequestHandlers.ts
@@ -1,8 +1,8 @@
 import { NextFunction, Request, Response } from 'express'
-import logger from '../../../logger'
+import logger from '../../logger'
 
 /**
- * A module exporting request handler functions to support ensuring the Create A Goal process has been followed
+ * A module exporting request handler functions to support ensuring page requests have been followed
  * in the correct sequence - IE. pages are not being called out of sequence.
  */
 
@@ -81,9 +81,25 @@ const checkPrisonerSummaryExistsInSession = async (req: Request, res: Response, 
   }
 }
 
+/**
+ * Request handler function to check the UpdateGoalForm exists in the session for the prisoner referenced in the
+ * request URL.
+ */
+const checkUpdateGoalFormExistsInSession = async (req: Request, res: Response, next: NextFunction) => {
+  if (!req.session.updateGoalForm) {
+    logger.warn(
+      `No UpdateGoalForm object in session - user attempting to navigate to path ${req.path} out of sequence. Redirecting to Overview page.`,
+    )
+    res.redirect(`/plan/${req.params.prisonNumber}/view/overview`)
+  } else {
+    next()
+  }
+}
+
 export {
   checkCreateGoalFormExistsInSession,
   checkAddStepFormsArrayExistsInSession,
   checkAddNoteFormExistsInSession,
   checkPrisonerSummaryExistsInSession,
+  checkUpdateGoalFormExistsInSession,
 }
