--- server/routes/updateGoal/routerRequestHandlers.ts
@@ -1,24 +1,0 @@
-import { NextFunction, Request, Response } from 'express'
-import logger from '../../../logger'
-
-/**
- * A module exporting request handler functions to support ensuring the Update a goal process has been followed
- * in the correct sequence - IE. pages are not being called out of sequence.
- */
-
-/**
- * Request handler function to check the UpdateGoalForm exists in the session for the prisoner referenced in the
- * request URL.
- */
-const checkUpdateGoalFormExistsInSession = async (req: Request, res: Response, next: NextFunction) => {
-  if (!req.session.updateGoalForm) {
-    logger.warn(
-      `No UpdateGoalForm object in session - user attempting to navigate to path ${req.path} out of sequence. Redirecting to Overview page.`,
-    )
-    res.redirect(`/plan/${req.params.prisonNumber}/view/overview`)
-  } else {
-    next()
-  }
-}
-
-export default checkUpdateGoalFormExistsInSession
